requirements.txt
flask
fpdf
psycopg2-binary